// frontend/exam-generation-app/src/app/models/question-type-configuration.ts
export class QuestionTypeConfiguration {
    type: string; // Type of question (e.g., singleChoice, multipleChoice)
    easy: number; // Number of easy questions
    medium: number; // Number of medium questions
    hard: number; // Number of hard questions
    total: number; // Total number of questions
    isEditing?: boolean;

    constructor(type: string, easy: number, medium: number, hard: number, total: number) {
        this.type = type;
        this.easy = easy;
        this.medium = medium;
        this.hard = hard;
        this.total = total;
    }
}